WIP new slice migration
