+++ src/store/slices/app.ts
@@ -255,12 +255,9 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
     set({
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks: {
-            ...state.data.deckCollection.decks,
-            [deck.id]: deck,
-          },
+        decks: {
+          ...state.data.decks,
+          [deck.id]: deck,
         },
         history: {
           ...state.data.history,
@@ -274,7 +271,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
   },
   async deleteDeck(id, cb) {
     const state = get();
-    const decks = { ...state.data.deckCollection.decks };
+    const decks = { ...state.data.decks };
 
     const deckEdits = { ...state.deckEdits };
     delete deckEdits[id];
@@ -305,10 +302,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
     set({
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks,
-        },
+        decks,
         history,
       },
       deckEdits,
@@ -318,7 +312,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
   async deleteAllDecks() {
     const state = get();
 
-    const decks = { ...state.data.deckCollection.decks };
+    const decks = { ...state.data.decks };
     const history = { ...state.data.history };
     const edits = { ...state.deckEdits };
 
@@ -333,10 +327,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
     set({
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks,
-        },
+        decks,
         history,
       },
     });
@@ -352,7 +343,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
 
     const edits = state.deckEdits[deckId];
 
-    const deck = state.data.deckCollection.decks[deckId];
+    const deck = state.data.decks[deckId];
     if (!deck) return deckId;
 
     const nextDeck = applyDeckEdits(deck, edits, state.metadata, true);
@@ -377,12 +368,9 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
       deckEdits,
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks: {
-            ...state.data.deckCollection.decks,
-            [nextDeck.id]: nextDeck,
-          },
+        decks: {
+          ...state.data.decks,
+          [nextDeck.id]: nextDeck,
         },
       },
     });
@@ -395,7 +383,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
 
     const state = get();
 
-    const deck = state.data.deckCollection.decks[id];
+    const deck = state.data.decks[id];
     assert(deck, `Deck ${id} does not exist.`);
 
     const now = new Date().toISOString();
@@ -462,13 +450,10 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
       deckEdits,
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks: {
-            ...state.data.deckCollection.decks,
-            [deck.id]: prevDeck,
-            [newDeck.id]: newDeck,
-          },
+        decks: {
+          ...state.data.decks,
+          [deck.id]: prevDeck,
+          [newDeck.id]: newDeck,
         },
         history,
       },
@@ -480,14 +465,14 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
   async deleteUpgrade(id, cb) {
     const state = get();
 
-    const deck = state.data.deckCollection.decks[id];
+    const deck = state.data.decks[id];
     assert(deck, `Deck ${id} does not exist.`);
 
     const previousId = deck.previous_deck;
 
     assert(previousId, "Deck does not have a previous deck.");
 
-    const decks = { ...state.data.deckCollection.decks };
+    const decks = { ...state.data.decks };
     assert(decks[previousId], "Previous deck does not exist.");
 
     const history = { ...state.data.history };
@@ -511,10 +496,7 @@ export const createAppSlice: StateCreator<StoreState, [], [], AppSlice> = (
       deckEdits,
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks,
-        },
+        decks,
         history,
       },
     });
