+++ src/store/slices/data.ts
@@ -18,12 +18,7 @@ import { type DataSlice, type Deck, type Id, isDeck } from "./data.types";
 export function getInitialDataState() {
   return {
     data: {
-      deckCollection: {
-        filters: {
-          faction: [],
-        },
-        decks: {},
-      },
+      decks: {},
       history: {},
     },
   };
@@ -45,12 +40,9 @@ export const createDataSlice: StateCreator<StoreState, [], [], DataSlice> = (
     set({
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks: {
-            ...state.data.deckCollection.decks,
-            [deck.id]: deck,
-          },
+        decks: {
+          ...state.data.decks,
+          [deck.id]: deck,
         },
         history: {
           ...state.data.history,
@@ -81,15 +73,12 @@ export const createDataSlice: StateCreator<StoreState, [], [], DataSlice> = (
     set({
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks: {
-            ...state.data.deckCollection.decks,
-            ...decks.reduce<Record<Id, Deck>>((acc, deck) => {
-              acc[deck.id] = deck;
-              return acc;
-            }, {}),
-          },
+        decks: {
+          ...state.data.decks,
+          ...decks.reduce<Record<Id, Deck>>((acc, deck) => {
+            acc[deck.id] = deck;
+            return acc;
+          }, {}),
         },
         history: {
           ...state.data.history,
@@ -105,7 +94,7 @@ export const createDataSlice: StateCreator<StoreState, [], [], DataSlice> = (
   duplicateDeck(id) {
     const state = get();
 
-    const deck = state.data.deckCollection.decks[id];
+    const deck = state.data.decks[id];
     assert(deck, `Deck ${id} does not exist.`);
 
     const now = new Date().toISOString();
@@ -126,12 +115,9 @@ export const createDataSlice: StateCreator<StoreState, [], [], DataSlice> = (
     set({
       data: {
         ...state.data,
-        deckCollection: {
-          filters: { ...state.data.deckCollection.filters },
-          decks: {
-            ...state.data.deckCollection.decks,
-            [newDeck.id]: newDeck,
-          },
+        decks: {
+          ...state.data.decks,
+          [newDeck.id]: newDeck,
         },
         history: {
           ...state.data.history,
@@ -146,7 +132,7 @@ export const createDataSlice: StateCreator<StoreState, [], [], DataSlice> = (
   exportJSON(id) {
     const state = get();
 
-    const deck = state.data.deckCollection.decks[id];
+    const deck = state.data.decks[id];
     assert(deck, `Deck ${id} does not exist.`);
 
     const deckExport = formatDeckExport(deck);
@@ -168,29 +154,5 @@ export const createDataSlice: StateCreator<StoreState, [], [], DataSlice> = (
       `arkhambuild-${deck.id}.md`,
       "text/markdown",
     );
-  },
-  addDecksFilter(type, value) {
-    const state = get();
-    const filterValues = structuredClone(state.data.deckCollection.filters);
-
-    switch (type) {
-      case "faction":
-        filterValues[type] = value;
-    }
-
-    set({
-      data: {
-        ...state.data,
-        deckCollection: {
-          filters: filterValues,
-          decks: {
-            ...state.data.deckCollection.decks,
-          },
-        },
-        history: {
-          ...state.data.history,
-        },
-      },
-    });
   },
 });
